+++ Assets/Scripts/Interfaces/IHealth.cs
@@ -1,6 +1,6 @@
 namespace Interfaces {
     public interface IHealth {
-        public void TakeDamages(uint damage);
-        public void Heal(uint heal);
+        public void TakeDamages(ushort damage);
+        public void Heal(ushort heal);
     }
 }
